requirements.txt
openai==1.3.8
python-dotenv==1.0.0
langchain==0.1.1
langchain-openai==0.0.2
flask==3.0.0
requests==2.31.0
pydantic==2.5.0
